Add CSS enhancement to sortable tables
this improves on lucee#893 
https://luceeserver.atlassian.net/browse/LDEV-2777

improvement by adding cursor:pointer and down arrow as CSS to the headers of the sortable tables 'Execution Time' and 'Implicit Variable Access', to make more clear and suggest that these are clickable headers.

This PR is a replacement for PR lucee#946 that had a CLA Signing issue. Please excuse the inconvenience caused.