Tests for string, number, true, false & null
According to the JSON specification, simple top-level values consisting on javascript escaped strings, numbers, `true`, `false` and `null` are considered valid JSON.  Added tests to inspect the deserialized item to ensure it's what is expected. Added UTF-8 emoji test.

A third-party recently submitted a double-encoded JSON string and, while isJson() evaluated `true`, the value wasn't useable as it was a simple string instead of the expected object/array.  (I'm considering writing a `isJsonStucture()` user-defined function to perform additional checks to ensure that the parsed value is an object or array.)