LDEV-4745 address issue where classes are not fetched from class loader when they already exist
This appears to arise when the PageSource for a page is flushed/cleared but when the class
still exists in the class loader and using the latest version of the source.

Fix works by checking whether we already have a compiled binary class file that is newer than
source file and attempting to fetch the class from the class loader.